refactor(cli): extract sandbox connect action
