+++ src/lib/sandbox-runtime-actions.ts
@@ -3,14 +3,17 @@
 
 /* v8 ignore start -- transitional action facade until implementations leave src/nemoclaw.ts. */
 
-import type { SandboxConnectOptions } from "./nemoclaw-runtime-bridge";
+import type { SandboxConnectOptions } from "./sandbox-connect-action";
 import { getNemoClawRuntimeBridge } from "./nemoclaw-runtime-bridge";
 
 export async function connectSandbox(
   sandboxName: string,
   options?: SandboxConnectOptions,
 ): Promise<void> {
-  await getNemoClawRuntimeBridge().sandboxConnect(sandboxName, options);
+  const { connectSandbox: connectExtractedSandbox } = require("./sandbox-connect-action") as {
+    connectSandbox: (sandboxName: string, options?: SandboxConnectOptions) => Promise<void>;
+  };
+  await connectExtractedSandbox(sandboxName, options);
 }
 
 export async function showSandboxStatus(sandboxName: string): Promise<void> {
